refactor(frontend): one type per file in /types/entities
