+++ apps/frontend/src/types/entities/base.entity.ts
@@ -1,0 +1,10 @@
+export abstract class Base {
+    id: string;
+    createdAt: Date;
+    updatedAt: Date;
+    deletedAt: Date;
+  
+    constructor(values: Partial<Base>) {
+      Object.assign(this, values);
+    }
+  }
