+++ apps/frontend/src/types/entities/projects/project-action.enum.ts
@@ -1,0 +1,6 @@
+export enum ProjectAction {
+    read = 'project__read',
+    update = 'project__update',
+    create = 'project__create',
+    manage = 'project__manage',
+  }
