+++ apps/frontend/src/types/entities/subject.entity.ts
@@ -1,0 +1,5 @@
+import { Base } from "./base.entity";
+
+export class Subject extends Base {
+    public static actions: Record<string, string>;
+  }
