+++ apps/frontend/src/types/entities/users/user-status.enum.ts
@@ -1,0 +1,4 @@
+export enum UserStatus {
+  active = 0,
+  pending_email_activation = 1,
+}
